Update App Extension bindings.
