--- src/AirshipBindings.iOS.AppExtensions/AirshipExtensionsStructs.cs
@@ -1,9 +1,0 @@
-﻿/*
- Copyright Airship and Contributors
-*/
-
-using System;
-
-namespace AirshipBindings.iOS.AppExtensions
-{
-}
