+++ src/AirshipBindings.iOS.AppExtensions/StructsAndEnums.cs
@@ -1,0 +1,11 @@
+/*
+ Copyright Airship and Contributors
+*/
+using ObjCRuntime;
+using System;
+using UIKit;
+
+namespace AirshipBindings.iOS.AppExtensions {
+
+
+}
